add regex

+++ src/main/java/com/ua/sutty/servlets/DeleteServlet.java
@@ -1,5 +1,6 @@
 package com.ua.sutty.servlets;
 
+import com.ua.sutty.domain.User;
 import com.ua.sutty.repository.impl.JdbcUserDao;
 import com.ua.sutty.utils.DataSource;
 
@@ -15,29 +16,29 @@ public class DeleteServlet extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.setAttribute("successfullDelete", 2);
+        User user = (User) req.getSession().getAttribute("loggedInUser");
         JdbcUserDao jdbcUserDao = new JdbcUserDao(new DataSource().getBasicDataSourceTest());
-
         String id = req.getParameter("userIdForDelete");
-
-        System.out.println("Don't have id" + id);
-        if (req.getParameter("successfullDelete") == null || Integer.valueOf(req.getParameter("successfullDelete")) != 1 ) {
+        if (req.getParameter("successfullyDeleted") == null || Integer.valueOf(req.getParameter("successfullyDeleted")) != 1) {
             if (id == null) {
                 req.getServletContext().getRequestDispatcher("/jsp/errorPage.jsp").forward(req, resp);
                 return;
             }
         }
+        if (user.getId().equals(Long.valueOf(id))) {
+            req.getSession().setAttribute("errorDeleting", true);
+            resp.sendRedirect("/home");
+            return;
+        }
         jdbcUserDao.remove(jdbcUserDao.findById(Long.valueOf(id)));
-        req.getSession().setAttribute("successfullDelete", 1);
-//        req.setAttribute("successfullDelete", 1);
+        req.getSession().setAttribute("successfullyDeleted", true);
         req.getSession().setAttribute("users", jdbcUserDao.findAll());
-//        req.getServletContext().getRequestDispatcher("/jsp/admin-home.jsp").forward(req, resp);
         resp.sendRedirect("/home");
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-
+        doGet(req, resp);
     }
 
 }
